Fixes and changes (WIP)
Renamed index to diff_config, as index is confusing.
Registration needs to be exactly the same for Imag and Real, otherwise we will get nasty artefacts. So we need to apply a transform instead of running the registration.
Changed the debug rigid registration to a mean of the stack before and after registration.