Fixes HA line profiles for phantom images
Addresses an issue where NaN values in the average HA line profile,
particularly when processing phantom images, caused errors during linear
regression fitting.

The changes add a check for NaNs in the average HA line profile. If
NaNs are found, the linear regression fitting is skipped for that
slice, preventing errors and ensuring the rest of the data can still
be processed. Also add logger as argument to function, to enable
debug messages.