Fixes log file handling and montage creation
Ensures proper handling of log files when re-running analysis
in the same folder by closing the existing file handler
before assigning a new one.

Adds functionality to create a 2D montage of accepted DWIs
even when no outlier removal is performed. This ensures that
the results are always visualized, even when all images are accepted.